+++ packages/core-js/full/url/can-parse.js
@@ -1,0 +1,3 @@
+var parent = require('../../actual/url/can-parse');
+
+module.exports = parent;