+++ packages/core-js/proposals/url.js
@@ -1,5 +1,6 @@
 // https://github.com/jasnell/proposal-url
 require('../modules/web.url');
+require('../modules/web.url.can-parse');
 require('../modules/web.url.to-json');
 require('../modules/web.url-search-params');
 require('../modules/web.url-search-params.size');