+++ packages/core-js/stable/url/can-parse.js
@@ -1,0 +1,5 @@
+require('../../modules/web.url');
+require('../../modules/web.url.can-parse');
+var path = require('../../internals/path');
+
+module.exports = path.URL.canParse;