+++ packages/core-js/web/index.js
@@ -11,6 +11,7 @@ require('../modules/web.self');
 require('../modules/web.structured-clone');
 require('../modules/web.timers');
 require('../modules/web.url');
+require('../modules/web.url.can-parse');
 require('../modules/web.url.to-json');
 require('../modules/web.url-search-params');
 require('../modules/web.url-search-params.size');