+++ tests/compat/tests.js
@@ -1936,6 +1936,9 @@ GLOBAL.tests = {
   // TODO: Remove this module from `core-js@4` since it's split to submodules
   'web.timers': TIMERS,
   'web.url.constructor': URL_AND_URL_SEARCH_PARAMS_SUPPORT,
+  'web.url.can-parse': [URL_AND_URL_SEARCH_PARAMS_SUPPORT, function () {
+    return URL.canParse;
+  }],
   'web.url.to-json': [URL_AND_URL_SEARCH_PARAMS_SUPPORT, function () {
     return URL.prototype.toJSON;
   }],